+++ change_log.md
@@ -1,6 +1,22 @@
 # OWASP Java HTML Sanitizer Change Log
 
 Most recent at top.
+  * Release 20190610.1
+    * Recognize named HTML entities added in the last few years.
+  * Release 20190503.1
+    * Make Encoding class public so that clients can use HTML text decoder.
+    * Fix bug in srcset handling.
+  * Release 20190325.1
+    * Properly parse `srcset` attribute values to apply URL policy to
+      each URL in turn.
+    * Update dependency on guava version to 27.1-jre to avoid causing clients
+      problems with CVE-2018-10237.  Specify Maven property `guava.version`
+      to override.
+    * Compatible with JDK 11.
+  * Release 20181114.1
+    * Compatible with guava > 19 including 21.x.x - 27.x.x
+    * Public API now supports custom style creation.
+    * Tweaks to handline of [HTML comment like constructs](https://www.w3.org/TR/html5/scripting-1.html#restrictions-for-contents-of-script-elements) in script element bodies.
   * Release 20180219.1
     * Strip ZWNJ from MacOS and iOS [crashing text sequences](https://manishearth.github.io/blog/2018/02/15/picking-apart-the-crashing-ios-string/)
   * Release 20171016.1